+++ .gitignore
@@ -36,3 +36,6 @@ next-env.d.ts
 
 # eslint
 .eslintcache
+
+# ide settings
+.idea
